+++ src/parsley/form.js
@@ -10,17 +10,18 @@ define('parsley/form', [
       throw new Error('You must give a Parsley instance');
 
     this.parsleyInstance = parsleyInstance;
-    return this.init($(element));
+    this.$element = $(element);
   };
 
   ParsleyForm.prototype = {
-    init: function ($element) {
-      this.$element = $element;
+    init: function () {
       this.validationResult = null;
 
       this.options = this.parsleyInstance.OptionsFactory.get(this);
 
-      return this._bindFields();
+      this._bindFields();
+
+      return this;
     },
 
     onSubmitValidate: function (event) {
@@ -91,8 +92,8 @@ define('parsley/form', [
       this.$element.find(this.options.inputs).each(function () {
         var fieldInstance = new window.Parsley(this, {}, self.parsleyInstance);
 
-        // only add valid field children
-        if ('ParsleyField' === fieldInstance.__class__)
+        // Only add valid and not excluded ParsleyField children
+        if ('ParsleyField' === fieldInstance.__class__  && !fieldInstance.$element.is(fieldInstance.options.excluded))
           self.fields.push(fieldInstance);
       });