+++ test/features/form.js
@@ -14,7 +14,7 @@ define(function () {
             '<div id="field2"></div>'           +
             '<textarea id="field2"></textarea>' +
           '</form>');
-        parsleyForm = new Parsley($('#element'));
+        var parsleyForm = new Parsley($('#element'));
         expect(parsleyForm.fields.length).to.be(2);
       });
       it('should bind parsleyFields children, and not excluded ones', function () {
@@ -24,9 +24,10 @@ define(function () {
             '<div id="field2"></div>'           +
             '<textarea id="field2"></textarea>' +
             '<div data-parsley-validate></div>' + // ParsleyForm, not a valid child
+            '<input id="field3" disabled />'    + // Disabled, excluded buy custom options below
             '<input type="submit"/>'            + // Excluded field, not valid
           '</form>');
-        parsleyForm = new Parsley($('#element'));
+        var parsleyForm = new Parsley($('#element'), { excluded: '[disabled], input[type=button], input[type=submit], input[type=reset]' });
         expect(parsleyForm.fields.length).to.be(2);
       });
       it('should properly bind options for form and children fields', function () {
@@ -36,7 +37,7 @@ define(function () {
             '<div id="field2"></div>'                                         +
             '<textarea id="field3" data-parsley-notblank="true"></textarea>'  +
           '</form>');
-        parsleyForm = new Parsley($('#element'));
+        var parsleyForm = new Parsley($('#element'));
         expect(parsleyForm.fields.length).to.be(2);
         expect(new Parsley('#field1').options.trigger).to.be('change');
         expect(new Parsley('#field1').options).to.have.key('required');
@@ -51,7 +52,7 @@ define(function () {
             '<div id="field2"></div>'                                         +
             '<textarea id="field3" data-parsley-notblank="true"></textarea>'  +
           '</form>');
-          parsleyForm = new Parsley($('#element'));
+          var parsleyForm = new Parsley($('#element'));
           parsleyForm.validate();
           expect(parsleyForm.validationResult).to.be(false);
           $('#field1').val('foo');
@@ -65,7 +66,7 @@ define(function () {
             '<div id="field2"></div>'                                                                  +
             '<textarea id="field3" data-parsley-group="bar" data-parsley-required="true"></textarea>'  +
           '</form>');
-          parsleyForm = new Parsley($('#element'));
+          var parsleyForm = new Parsley($('#element'));
           expect(parsleyForm.isValid()).to.be(false);
           $('#field1').val('value');
           expect(parsleyForm.isValid()).to.be(false);
@@ -79,18 +80,34 @@ define(function () {
             '<div id="field2"></div>'                                         +
             '<textarea id="field3" data-parsley-notblank="true"></textarea>'  +
           '</form>');
-          parsleyForm = new Parsley($('#element'));
+          var parsleyForm = new Parsley($('#element'));
+
+          // parsley.remote hack because if valid, parsley remote re-send form
+          parsleyForm.subscribe('parsley:form:validate', function (formInstance) {
+            if (formInstance.asyncSupport)
+              formInstance.submitEvent._originalPreventDefault();
+          });
+
           var event = $.Event();
+          // parsley.remote hack
+          event._originalPreventDefault = event.preventDefault;
           event.preventDefault = sinon.spy();
           parsleyForm.onSubmitValidate(event);
           expect(event.preventDefault.called).to.be(true);
 
           $('#field1').val('foo');
           $('#field3').val('foo');
+
           event = $.Event();
+          // parsley.remote hack
+          event._originalPreventDefault = event.preventDefault;
           event.preventDefault = sinon.spy();
           parsleyForm.onSubmitValidate(event);
-          expect(event.preventDefault.called).to.be(false);
+
+          if (!parsleyForm.asyncSupport)
+            expect(event.preventDefault.called).to.be(false);
+          else
+            expect(event.preventDefault.called).to.be(true);
       });
       it('should have a force option for validate and isValid methods', function () {
         $('body').append(